# api/swagger.yml
openapi: 3.1.0
info:
  title: Receipt Wrangler API.
  termsOfService:
    there are no TOS at this moment, use at your own risk we take no
    responsibility
  license:
    name: GPLV3
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
  version: 5.0.0
servers:
  - url: /api
security:
  - bearerAuth: [ ]
  - apiKeyAuth: [ ]
paths:
...
  /systemSettings:
    get:
      tags:
        - SystemSettings
      summary: Get system settings
      description: This will get system settings
      operationId: getSystemSettings
      responses:
        200:
          description: The system settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemSettings"
        500:
          $ref: "#/components/responses/Internal"
        403:
          $ref: "#/components/responses/Forbidden"
    put:
      tags:
        - SystemSettings
      summary: Update system settings
      description: This will update system settings
      operationId: updateSystemSettings
      requestBody:
        description: System settings to update
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertSystemSettingsCommand"
      responses:
        200:
          description: The updated system settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemSettings"
        500:
          $ref: "#/components/responses/Internal"
        403:
          $ref: "#/components/responses/Forbidden"
  /systemSettings/restartTaskServer:
    post:
      tags:
        - SystemSettings
      summary: Restart task server
      description: This will restart the task server
      operationId: restartTaskServer
      responses:
        200:
          $ref: "#/components/responses/Ok"
        500:
          $ref: "#/components/responses/Internal"
        403:
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: API Key in format v1.xxx.xxx.xxx
  responses:
    Ok:
      description: Request was successfully processed
...
    Forbidden:
      description: The request was not allowed
...
    Internal:
      description: There was an error processing the request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalErrorResponse'
  schemas:
...
    CurrencySeparator:
      type: string
      enum:
        - ","
        - "."
    CurrencySymbolPosition:
      type: string
      enum:
        - "START"
        - "END"
    QueueName:
      type: string
      enum:
        - "quick_scan"
        - "email_polling"
        - "email_receipt_processing"
        - "email_receipt_image_cleanup"
...
    BaseModel:
      type: object
      required:
        - id
        - createdAt
      properties:
        id:
          type: integer
          format: uint64
        createdAt:
          type: string
        createdBy:
          type: integer
          format: uint64
          default: 0
        createdByString:
          type: string
          description: Created by entity's name
          default: ""
        updatedAt:
          type: string
          default: ""
...
    TaskQueueConfiguration:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - type: object
          properties:
            name:
              $ref: "#/components/schemas/QueueName"
            priority:
              type: integer
              description: Queue priority
    UpsertTaskQueueConfiguration:
      type: object
      properties:
        name:
          $ref: "#/components/schemas/QueueName"
        priority:
          type: integer
          description: Queue priority
    SystemSettings:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - type: object
          required:
            - taskQueueConfigurations
          properties:
            enableLocalSignUp:
              type: boolean
              description: Whether local sign up is enabled
              default: false
            currencyDisplay:
              type: string
              description: Currency display
              default: "$"
            currencyThousandthsSeparator:
              $ref: "#/components/schemas/CurrencySeparator"
            currencyDecimalSeparator:
              $ref: "#/components/schemas/CurrencySeparator"
            currencySymbolPosition:
              $ref: "#/components/schemas/CurrencySymbolPosition"
            currencyHideDecimalPlaces:
              type: boolean
              description: Whether to hide decimal places
              default: false
            debugOcr:
              type: boolean
              description: Debug OCR
              default: false
            numWorkers:
              type: integer
              description: Number of workers to use
              default: 1
            emailPollingInterval:
              type: integer
              description: Email polling interval
              default: 1800
            receiptProcessingSettingsId:
              type: integer
              description: Receipt processing settings foreign key
            fallbackReceiptProcessingSettingsId:
              type: integer
              description: Fallback receipt processing settings foreign key
            taskConcurrency:
              type: integer
              description: Concurrency for task worker
              default: 10
            taskQueueConfigurations:
              type: array
              items:
                $ref: "#/components/schemas/TaskQueueConfiguration"
    UpsertSystemSettingsCommand:
      type: object
      required:
        - currencyThousandthsSeparator
        - currencyDecimalSeparator
        - currencyHideDecimalPlaces
        - currencySymbolPosition
        - taskConcurrency
      properties:
        enableLocalSignUp:
          type: boolean
          description: Whether local sign up is enabled
        currencyDisplay:
          type: string
          description: Currency display
        currencyThousandthsSeparator:
          $ref: "#/components/schemas/CurrencySeparator"
        currencyDecimalSeparator:
          $ref: "#/components/schemas/CurrencySeparator"
        currencySymbolPosition:
          $ref: "#/components/schemas/CurrencySymbolPosition"
        currencyHideDecimalPlaces:
          type: boolean
          description: Whether to hide decimal places
        debugOcr:
          type: boolean
        numWorkers:
          type: integer
          description: Number of workers to use
          default: 1
        emailPollingInterval:
          type: integer
          description: Email polling interval
        receiptProcessingSettingsId:
          type: integer
          description: Receipt processing settings foreign key
        fallbackReceiptProcessingSettingsId:
          type: integer
          description: Fallback receipt processing settings foreign key
        taskConcurrency:
          type: integer
          description: Concurrency for task worker
        taskQueueConfigurations:
          type: array
          items:
            $ref: "#/components/schemas/UpsertTaskQueueConfiguration"
...
    InternalErrorResponse:
      type: object
      required:
        - errorMsg
      properties:
        errorMsg:
          type: string
          description: Error message
...